refactor: seal `Field`s
