+++ packages/widgetbook/lib/src/fields/fields.dart
@@ -1,13 +1,3 @@
-export 'boolean_field.dart';
-export 'color_field/color_field.dart';
-export 'date_time_field.dart';
-export 'double_input_field.dart';
-export 'double_slider_field.dart';
-export 'duration_field.dart';
+export 'color_field/color_space.dart';
 export 'field.dart';
 export 'fields_composable.dart';
-export 'int_input_field.dart';
-export 'int_slider_field.dart';
-export 'object_dropdown_field.dart';
-export 'object_segmented_field.dart';
-export 'string_field.dart';
