+++ packages/widgetbook/lib/src/fields/int_slider_field.dart
@@ -1,10 +1,7 @@
-import 'package:flutter/material.dart';
-
-import 'field.dart';
-import 'num_slider_field.dart';
+part of 'field.dart';
 
 /// A [Field] that builds [Slider] for [int] values.
-class IntSliderField extends NumSliderField<int> {
+final class IntSliderField extends NumSliderField<int> {
   /// Creates a new instance of [IntSliderField].
   IntSliderField({
     required super.name,
