+++ packages/widgetbook/lib/widgetbook.dart
@@ -10,7 +10,7 @@ export 'package:flutter_test/src/widget_tester.dart';
 export 'src/addons/addons.dart' hide NoneViewport;
 export 'src/args/args.dart';
 export 'src/core/core.dart';
-export 'src/fields/fields.dart' hide DateTimeExtension;
+export 'src/fields/fields.dart';
 export 'src/routing/query_group.dart';
 export 'src/runner.dart';
 export 'src/state/state.dart' hide WidgetbookScope;
